Prometheus doesn't like custom timestamps
Originally, I tried to use the timestamp field to track the date given
by homebrew for the metrics. Prometheus complained that the metrics
were too far out of the appropriate window, so instead add the date
given as a label and let prometheus set it's own timestamp about when it
was scraped.